Create a generic bundle component.
Add missing bundles files.

Removed extra MUA layout spacing.
