+++ src/presentational-components/myUserAccess/bundles.js
@@ -28,6 +28,7 @@ export const bundleData = [
       'automation hub': '/automation-hub',
       'automation services catalog': '/catalog',
     },
+    appsIds: ['catalog', 'approval', 'automation-analytics', 'automation-hub'],
   },
   {
     entitlement: 'cost_management',
@@ -43,6 +44,7 @@ export const bundleData = [
     apps: {
       'migration analytics': '/migration-analytics',
     },
+    appsIds: ['migration-analytics'],
   },
   {
     entitlement: 'subscriptions',
@@ -51,5 +53,6 @@ export const bundleData = [
       'Red Hat Enterprise Linux': '/rhel-sw',
       'Red Hat OpenShift': '/openshift-sw',
     },
+    appsIds: ['subscriptions'],
   },
 ];
