+++ src/smart-components/myUserAccess/MUAHome.scss
@@ -15,12 +15,10 @@
         display: grid;
         grid-template-columns: 1fr 2fr;
         column-gap: 24px;
-        padding: 0 24px;
         background: white;
         .ins-l-myUserAccess-section .pf-c-title { margin-top: 15px; }
         .ins-l-myUserAccess-section:first-of-type {
             border-right: 1px solid var(--pf-global--BorderColor--100);
-            padding-right: 24px;
         }
     }
